package.json: Update dependencies
Update dependencies so that npm install is successful. We need to
be able to install our deps so we can build the project into dist,
as this is actually what we consume from npmjs.com, but it doesn't
exist when we use the project as a git dependency.

Now we need to add a `prepare` script so that npm runs it before
installing the package.

See: https://stackoverflow.com/a/57829251